Fix Java response schema
Put Java extra response data in a meta field (more generic, easier to write rule for).

It also fix the runTime field (now meta.runTime). It was being overwritten by runCount.

Because run time is unpredictable (and now shows), it is marked as ‘<ANY>’ in the e2e tests’ expectation values (see example/tests.yaml). I changed `bin/verifier test` command to skip value expected as ‘<ANY>’ as long as they are set.

The tests had to be tweaked to work with the new verifier test runner.